+++ src/main/java/dev/yatloaf/modkrowd/custom/SelfAlohaMessage.java
@@ -1,11 +1,11 @@
 package dev.yatloaf.modkrowd.custom;
 
+import dev.yatloaf.modkrowd.cubekrowd.common.CKColor;
 import dev.yatloaf.modkrowd.cubekrowd.common.cache.TextCache;
 import dev.yatloaf.modkrowd.cubekrowd.message.Aloha;
 import net.minecraft.client.network.ServerInfo;
 import net.minecraft.client.resource.language.I18n;
 import net.minecraft.text.Text;
-import net.minecraft.util.Formatting;
 
 public record SelfAlohaMessage(Aloha aloha, ServerInfo info) implements Custom {
     public String nameOrAddress() {
@@ -17,6 +17,6 @@ public record SelfAlohaMessage(Aloha aloha, ServerInfo info) implements Custom {
         return TextCache.of(Text.translatable(
                 this.aloha == Aloha.JOIN ? "modkrowd.message.self_joined" : "modkrowd.message.self_left",
                 this.nameOrAddress()
-        ).formatted(Formatting.AQUA));
+        ).setStyle(CKColor.AQUA.style));
     }
 }